version-a-durable-functions/requirements.txt
azure-functions>=1.21.0
azure-functions-durable>=1.2.9
